runtime/python/agentschema/pyproject.toml
[project]
name = "AgentSchema"
version = "0.1.0"
description = "Declarative Agent Schema"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "pyyaml>=6.0.1",
    "python-dotenv>=1.0.1",
    "click>=8.1.7",
    "aiofiles>=24.1.0",
    "ruff>=0.13.3",
    "black>=25.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "ruff",
    "mypy",
    "types-aiofiles",
    "types-PyYAML",
    "pytest-cov",
    "black",
]

[build-system]
requires = ["uv_build>=0.8.23,<0.9.0"]
build-backend = "uv_build"

[tool.uv.workspace]
members = ["."]

[tool.uv.sources]
agentschema = { workspace = true }

[dependency-groups]
dev = ["agentschema", "pytest>=8.4.2"]
